# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "DASH-IF provides conformance software, a set of test vectors, and an open source reference client implementation for DASH-AVC/264. If you are considering a DASH deployment, then a DASH-AVC/264 compatible solution is the best place for you to start\n\n\n#### Disclaimer\n\n\nThese test vectors are being made available under the Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International license. DASH-IF considers the use of these test vectors for the development of commercial DASH clients and services to be a non-commercial use under the Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International license provided that the test vectors themselves are not distributed for commercial purposes.\n\n\n### Available Content\nIssue Reporting. If any issue is noticed in the test vectors, please file an issue report [here](https://github.com/Dash-Industry-Forum/Test-Vectors/issues) , including the test vector link and an informal description of the issue (GitHub user account required for creating issue).\n\n\nTest vectors may be downloaded as a group via WGET by recursing this top level directory: http://dash.edgesuite.net/dash264/TestCases/\n\n\n[Learn more about DASH-IF](http://www.dashif.org)\n",
    "version": "1.0.0",
    "title": "DASH-IF Test Vectors API",
    "termsOfService": "TBD",
    "contact": {
      "name": "TBD"
    },
    "license": {
      "name": "TBD",
      "url": "TBD"
    }
  },
  "host": "192.168.99.100:3000",
  "basePath": "/v1",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "description": "Gets a list of existing DASH-IF Testvector system users who can create / delete `Testvector` and `Testcontent`. Only user with administrator role can obtain this list of users.\n",
        "security": [
          {
            "Bearer": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful responses",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "get": {
        "description": "Authorize User by username and password and if User is authorized succesffully issue an authorization JWT (JSON Web Token)\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/TBD_AuthJWT"
            }
          },
          "default": {
            "description": "TBD (prop. unexpected error)",
            "schema": {
              "$ref": "#/definitions/TBD_ErrorModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "companyname": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "createdAt": {
          "type": "date"
        },
        "updatedAt": {
          "type": "date"
        }
      }
    },
...
  }
}